pyproject.toml
[tool.poetry]
name = "projectdrf"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.4"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.10"
djangorestframework = "^3.15.2"
django-filter = "^24.3"
djangorestframework-simplejwt = "^5.3.1"
drf-yasg = "^1.21.8"
django-cors-headers = "^4.6.0"
redis = "^5.2.1"
celery = "^5.4.0"
eventlet = "^0.39.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
